refactor : 매직 넘버 상수화 및 구조 수정
