+++ src/main/java/laddergameplay/view/ResultView.java
@@ -4,7 +4,7 @@ import laddergameplay.domain.data.Person;
 import laddergameplay.domain.data.Result;
 import laddergameplay.domain.datas.People;
 import laddergameplay.domain.datas.Results;
-import laddergameplay.domain.dto.WinningResult;
+import laddergameplay.domain.WinningResult;
 import laddergameplay.domain.ladder.Ladder;
 
 import java.util.Map;
